requirements.txt
Flask==0.10.1
nltk==3.2.1
lxml==3.6.4
numpy==1.11.2
biopython==1.68
beautifulsoup4==4.4.1
scikit-learn==0.18.1
scipy==0.17.0

